Refactor RecipeJson and add bulk recipe creation endpoint
This commit refactors the `RecipeJson`, `RecipeCategory`, `RecipeTag`, and `RecipeTool` data classes into their own files under `com.saintpatrck.mealie.client.api.model`.

It also introduces the `createFromUrlBulk` endpoint to the `RecipesApi`, allowing for the creation of multiple recipes from URLs in a single request. This includes:
- `CreateRecipeFromUrlBulkRequestJson` data class for the request body.
- `CreateRecipeFromUrlBulkResponseJson` data class for the response body.
- A corresponding test case `createRecipeFromUrlBulk should deserialize correctly` in `RecipesApiTest`.
